app/pyproject.toml
[tool.poetry]
name = "erfiume"
version = "0.1.0"
description = ""
authors = ["Edoardo Rosa <6991986+notdodo@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
aioboto3 = "^13.1.1"
python-telegram-bot = "^21.5"
poetry-dotenv-plugin = "^0.2.0"

[tool.poetry.group.dev.dependencies]
aws-lambda-powertools = "^2.43.1"
awscli-local = "^0.22.0"
boto3-stubs = { version = "1.35.5", extras = ["dynamodb", "secretsmanager"] }
mypy = "^1.10.1"
poetry-plugin-dotenv = "^2.4.0"
pydantic = "^2.9.2"
ruff = "^0.6.5"
types-aioboto3 = { extras = [
    "dynamodb",
    "secretsmanager",
], version = "^13.1.1" }

[tool.ruff]
output-format = "github"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D2", "D4", "ANN", "COM812", "ISC001"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
case-sensitive = true

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pycodestyle]
max-line-length = 150

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
pretty = true
show_error_codes = true
show_error_context = true
implicit_reexport = true

[tool.pylint."MESSAGES CONTROL"]
persistent = "no"
disable = ["fixme", "line-too-long"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
